:sparkles: [ADD] : first motion feedback plot simulation
this simulink file is used to plot the step response of the DC motor of the LMR.
since this attempt is solely purposed on plotting the step response at the initial stage the .slx file is not very detailed and the description written in the file is of enough help to understand the purpose and the blocks used. further adjustments and additions will be made in the simulink and simulation section later on.