📝 [DOC] : Investigating Motor Driver Design and Control
Attempted to identify the motor's dynamics equations based on its step response. However, it became apparent that the motor does not exhibit a linear characteristic. Notably, the motor reaches its ultimate speed much faster when the duty cycle is larger. This led to speculation about whether the non-linear input signal (theoretically pure PWM) or the motor itself (possibly due to the gearbox and extreme friction conditions) is responsible for this behavior.

To investigate further, we initially identified the system using a 100% duty cycle, which represents the motor's step response. Subsequently, we simulated the system with PWM input to determine if it accurately tracks the actual system with the same input signal. The results were intriguing, as the response exhibited significant differences in terms of the system's time constant. This suggests that the motor itself possesses a non-linear response.

Furthermore, our curiosity prompted us to examine the actual signals on the motors while employing an H-bridge motor driver. It is possible that the driver circuit also introduces non-linearity into the system. The results of the monitored signals have been included as attachments to this file.

Thus, we first identified the system with %100 duty cycle which is actually the step response of the motor. Then activate the system in simulation with PWM input to see if it tracks the actual system with the same input signal. The result was interesting since the response was significantly different in terms of system time constant. So it seems that the motor itself has a non-linear response.

Although, it raised our curiosity to see what is the actual signal on the motors while using a H-bridge motor driver. Maybe the driver circuit that add non-linearity to the system too. The results of the monitored signals are attached to this file.